+++ sdk/storage/azure-storage-file-share/src/samples/java/com/azure/storage/file/share/ShareAsyncJavaDocCodeSamples.java
@@ -5,6 +5,7 @@ package com.azure.storage.file.share;
 import com.azure.storage.common.StorageSharedKeyCredential;
 import com.azure.storage.file.share.models.ShareAccessPolicy;
 import com.azure.storage.file.share.models.ShareFileHttpHeaders;
+import com.azure.storage.file.share.models.ShareRequestConditions;
 import com.azure.storage.file.share.models.ShareSignedIdentifier;
 import com.azure.storage.file.share.models.NtfsFileAttributes;
 import com.azure.storage.file.share.sas.ShareSasPermission;
@@ -22,7 +23,7 @@ import java.util.Map;
  * Contains code snippets when generating javadocs through doclets for {@link ShareAsyncClient}.
  */
 public class ShareAsyncJavaDocCodeSamples {
-
+    private String leaseId = "leaseId";
 
     /**
      * Generates code sample for {@link ShareAsyncClient} instantiation.
@@ -242,6 +243,36 @@ public class ShareAsyncJavaDocCodeSamples {
     }
 
     /**
+     * Generates a code sample for using {@link ShareAsyncClient#createFileWithResponse(String, long, ShareFileHttpHeaders,
+     * FileSmbProperties, String, Map, ShareRequestConditions)}
+     */
+    public void createFileWithLease() {
+        ShareAsyncClient shareAsyncClient = createAsyncClientWithSASToken();
+        // BEGIN: com.azure.storage.file.share.ShareAsyncClient.createFileWithResponse#String-long-ShareFileHttpHeaders-FileSmbProperties-String-Map-ShareRequestConditions
+        ShareFileHttpHeaders httpHeaders = new ShareFileHttpHeaders()
+            .setContentType("text/html")
+            .setContentEncoding("gzip")
+            .setContentLanguage("en")
+            .setCacheControl("no-transform")
+            .setContentDisposition("attachment");
+        FileSmbProperties smbProperties = new FileSmbProperties()
+            .setNtfsFileAttributes(EnumSet.of(NtfsFileAttributes.READ_ONLY))
+            .setFileCreationTime(OffsetDateTime.now())
+            .setFileLastWriteTime(OffsetDateTime.now())
+            .setFilePermissionKey("filePermissionKey");
+        String filePermission = "filePermission";
+        // NOTE: filePermission and filePermissionKey should never be both set
+
+        ShareRequestConditions requestConditions = new ShareRequestConditions().setLeaseId(leaseId);
+
+        shareAsyncClient.createFileWithResponse("myfile", 1024, httpHeaders, smbProperties,
+            filePermission, Collections.singletonMap("directory", "metadata"), requestConditions)
+            .subscribe(response -> System.out.printf("Creating the file completed with status code %d",
+                response.getStatusCode()));
+        // END: com.azure.storage.file.share.ShareAsyncClient.createFileWithResponse#String-long-ShareFileHttpHeaders-FileSmbProperties-String-Map-ShareRequestConditions
+    }
+
+    /**
      * Generates a code sample for using {@link ShareAsyncClient#deleteDirectory(String)}
      */
     public void deleteDirectoryAsync() {
@@ -270,6 +301,23 @@ public class ShareAsyncJavaDocCodeSamples {
         );
         // END: com.azure.storage.file.share.ShareAsyncClient.deleteFile#string
     }
+
+    /**
+     * Generates a code sample for using {@link ShareAsyncClient#deleteFileWithResponse(String, ShareRequestConditions)}
+     */
+    public void deleteFileWithResponseAsync() {
+        ShareAsyncClient shareAsyncClient = createAsyncClientWithSASToken();
+        // BEGIN: com.azure.storage.file.share.ShareAsyncClient.deleteFile#string-ShareRequestConditions
+        ShareRequestConditions requestConditions = new ShareRequestConditions().setLeaseId(leaseId);
+        shareAsyncClient.deleteFileWithResponse("myfile", requestConditions).subscribe(
+            response -> {
+            },
+            error -> System.err.println(error.toString()),
+            () -> System.out.println("Complete deleting the file.")
+        );
+        // END: com.azure.storage.file.share.ShareAsyncClient.deleteFile#string-ShareRequestConditions
+    }
+
 
     /**
      * Generates a code sample for using {@link ShareAsyncClient#delete}