+++ sdk/storage/azure-storage-blob-cryptography/CHANGELOG.md
@@ -3,6 +3,10 @@
 ## 12.4.0-beta.1 (Unreleased)
 - Added support for upload via OutputStream by adding EncryptedBlobClient.getOutputStream methods
 
+## 12.3.1 (2020-02-10)
+- Updated `azure-core-http-netty` to version 1.3.0
+- Update `azure-storage-blob` to version 12.3.1
+
 ## 12.3.0 (2020-01-15)
 This package's
 [documentation](https://github.com/Azure/azure-sdk-for-java/blob/azure-storage-blob-cryptography_12.3.0/sdk/storage/azure-storage-blob-cryptography/README.md)